Change 'result' field to BooleanField
The intent of the 'result' field is to capture whether the transaction
resulted in a "match" or "no-match". The response from the provider
might include a lot of information, but the response has to filter down
to 'do we pass or fail the validation' which is what this field is
intended to capture, and so should be changed to reflect that.

Additionally, we want to allow null values in the database to handle the
case where we know a transaction has happened, but the result is either
unclear, or the transaction is not synchronous and we simply don't know
whether the transaction has succeeded or failed at this point.